First commit for GitHub
First off, the sensitive information in previous commits is no longer in use.
That info is now stored in special files that are ignored by git.

Added secret and mysql.cnf to .gitignore
Added a sample version of the mysql.cnf file
Updated file permissions on the majority of files
Added utility script files to run a test server or deploy the app to live